Start async rewrite (doesn't compile yet)
