+++ examples/gamepad/Cargo.toml
@@ -10,5 +10,5 @@ repository= "https://github.com/sciguy16/lego-powered-up"
 
 [dependencies]
 eyre = "0.6"
-gilrs = "0.8"
+gilrs = "0.10"
 lego-powered-up = { path = "../../lego-powered-up" }
