+++ lego-powered-up/Cargo.toml
@@ -13,15 +13,17 @@ readme = "../README.md"
 # https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-btleplug = "0.7"
+async-trait = "0.1"
+btleplug = "0.10"
 crossbeam-channel = "0.5"
-lazy_static = "1.4"
+lazy_static = "1"
 log = "0.4"
 lpu-macros = "0.1"
 num-derive = "0.3"
 num-traits = "0.2"
-thiserror = "1.0"
-uuid = "0.8"
+thiserror = "1"
+tokio = { version = "1", features = ["sync"] }
+uuid = "1"
 
 [dev-dependencies]
-env_logger = "0.9"
+env_logger = "0.10"
