+++ lego-powered-up/src/hubs.rs
@@ -5,14 +5,42 @@
 //! Specific implementations for each of the supported hubs.
 
 use crate::error::{OptionContext, Result};
-use crate::notifications::NotificationMessage;
-use crate::Hub;
-use crate::{
-    consts::blecharacteristic,
-    notifications::{AttachedIo, IoAttachEvent, VersionNumber},
-};
+// use crate::notifications::NotificationMessage;
+use crate::consts::blecharacteristic;
 use btleplug::api::{Characteristic, Peripheral, WriteType};
 use std::collections::HashMap;
+
+/// Trait describing a generic hub.
+#[async_trait::async_trait]
+pub trait Hub {
+    async fn name(&self) -> Result<String>;
+    async fn disconnect(&self) -> Result<()>;
+    async fn is_connected(&self) -> Result<bool>;
+    // The init function cannot be a trait method until we have GAT :(
+    //fn init(peripheral: P);
+    async fn properties(&self) -> &HubProperties;
+
+    async fn port_map(&self) -> &PortMap {
+        &self.properties().await.port_map
+    }
+
+    // cannot provide a default implementation without access to the
+    // Peripheral trait from here
+    async fn send_raw(&self, msg: &[u8]) -> Result<()>;
+
+    // fn send(&self, msg: NotificationMessage) -> Result<()>;
+
+    async fn subscribe(&self, char: Characteristic) -> Result<()>;
+
+    /// Ideally the vec should be sorted somehow
+    async fn attached_io(&self) -> Vec<ConnectedIo>;
+
+    // fn process_io_event(&mut self, _evt: AttachedIo);
+
+    async fn port(&self, port_id: Port) -> Result<PortController>;
+}
+
+pub type VersionNumber = u8;
 
 /// Propeties of a hub
 #[derive(Debug, Default)]
@@ -28,7 +56,7 @@ pub struct HubProperties {
     /// Battery level
     pub battery_level: usize,
     /// BLE signal strength
-    pub rssi: i8,
+    pub rssi: i16,
     /// Mapping from port type to port ID. Internally (to the hub) each
     /// port has a hardcoded identifier
     pub port_map: PortMap,
@@ -86,44 +114,52 @@ pub struct TechnicHub<P: Peripheral> {
     connected_io: HashMap<u8, ConnectedIo>,
 }
 
+#[async_trait::async_trait]
 impl<P: Peripheral> Hub for TechnicHub<P> {
-    fn name(&self) -> String {
-        self.peripheral.properties().local_name.unwrap_or_default()
+    async fn name(&self) -> Result<String> {
+        Ok(self
+            .peripheral
+            .properties()
+            .await?
+            .context("No properties found for hub")?
+            .local_name
+            .unwrap_or_default())
     }
 
-    fn disconnect(&self) -> Result<()> {
-        if self.is_connected() {
-            self.peripheral.disconnect()?;
+    async fn disconnect(&self) -> Result<()> {
+        if self.is_connected().await? {
+            self.peripheral.disconnect().await?;
         }
         Ok(())
     }
 
-    fn is_connected(&self) -> bool {
-        self.peripheral.is_connected()
+    async fn is_connected(&self) -> Result<bool> {
+        Ok(self.peripheral.is_connected().await?)
     }
 
-    fn properties(&self) -> &HubProperties {
+    async fn properties(&self) -> &HubProperties {
         &self.properties
     }
 
-    fn send_raw(&self, msg: &[u8]) -> Result<()> {
+    async fn send_raw(&self, msg: &[u8]) -> Result<()> {
         let write_type = WriteType::WithoutResponse;
         Ok(self
             .peripheral
-            .write(&self.lpf_characteristic, msg, write_type)?)
+            .write(&self.lpf_characteristic, msg, write_type)
+            .await?)
     }
 
-    fn send(&self, msg: NotificationMessage) -> Result<()> {
-        let msg = msg.serialise();
-        self.send_raw(&msg)?;
-        Ok(())
+    // fn send(&self, msg: NotificationMessage) -> Result<()> {
+    //     let msg = msg.serialise();
+    //     self.send_raw(&msg)?;
+    //     Ok(())
+    // }
+
+    async fn subscribe(&self, char: Characteristic) -> Result<()> {
+        Ok(self.peripheral.subscribe(&char).await?)
     }
 
-    fn subscribe(&self, char: Characteristic) -> Result<()> {
-        Ok(self.peripheral.subscribe(&char)?)
-    }
-
-    fn attached_io(&self) -> Vec<ConnectedIo> {
+    async fn attached_io(&self) -> Vec<ConnectedIo> {
         let mut ret = Vec::with_capacity(self.connected_io.len());
         for (_k, v) in self.connected_io.iter() {
             ret.push(v.clone());
@@ -134,31 +170,38 @@ impl<P: Peripheral> Hub for TechnicHub<P> {
         ret
     }
 
-    fn process_io_event(&mut self, evt: AttachedIo) {
-        match evt.event {
-            IoAttachEvent::AttachedIo { hw_rev, fw_rev } => {
-                if let Some(port) = self.port_from_id(evt.port) {
-                    let io = ConnectedIo {
-                        port_id: evt.port,
-                        port,
-                        fw_rev,
-                        hw_rev,
-                    };
-                    self.connected_io.insert(evt.port, io);
-                }
-            }
-            IoAttachEvent::DetachedIo { io_type_id: _ } => {}
-            IoAttachEvent::AttachedVirtualIo {
-                port_a: _,
-                port_b: _,
-            } => {}
-        }
+    // fn process_io_event(&mut self, evt: AttachedIo) {
+    //     match evt.event {
+    //         IoAttachEvent::AttachedIo { hw_rev, fw_rev } => {
+    //             if let Some(port) = self.port_from_id(evt.port) {
+    //                 let io = ConnectedIo {
+    //                     port_id: evt.port,
+    //                     port,
+    //                     fw_rev,
+    //                     hw_rev,
+    //                 };
+    //                 self.connected_io.insert(evt.port, io);
+    //             }
+    //         }
+    //         IoAttachEvent::DetachedIo { io_type_id: _ } => {}
+    //         IoAttachEvent::AttachedVirtualIo {
+    //             port_a: _,
+    //             port_b: _,
+    //         } => {}
+    //     }
+    // }
+
+    async fn port(&self, port_id: Port) -> Result<PortController> {
+        todo!()
     }
 }
 
 impl<P: Peripheral> TechnicHub<P> {
     /// Initialisation method
-    pub fn init(peripheral: P, chars: Vec<Characteristic>) -> Result<Self> {
+    pub async fn init(
+        peripheral: P,
+        chars: Vec<Characteristic>,
+    ) -> Result<Self> {
         // Peripheral is already connected before we get here
 
         println!("\n\nCHARACTERISTICS:\n\n{:?}\n\n", chars);
@@ -168,7 +211,10 @@ impl<P: Peripheral> TechnicHub<P> {
             .context("Device does not advertise LPF2_ALL characteristic")?
             .clone();
 
-        let props = peripheral.properties();
+        let props = peripheral
+            .properties()
+            .await?
+            .context("No properties found for hub")?;
 
         let mut port_map = PortMap::with_capacity(10);
         port_map.insert(Port::A, 0);
@@ -198,8 +244,8 @@ impl<P: Peripheral> TechnicHub<P> {
         })
     }
 
-    fn port_from_id(&self, port_id: u8) -> Option<Port> {
-        for (k, v) in self.port_map().iter() {
+    async fn port_from_id(&self, port_id: u8) -> Option<Port> {
+        for (k, v) in self.port_map().await.iter() {
             if *v == port_id {
                 return Some(*k);
             }
