+++ pu-util/Cargo.toml
@@ -10,7 +10,9 @@ repository = "https://github.com/sciguy16/lego-powered-up"
 
 [dependencies]
 anyhow = "1.0"
+async-trait = "0.1"
 clap = { version = "3", features = ["cargo"] }
-env_logger = "0.9"
+env_logger = "0.10"
 lego-powered-up = { path="../lego-powered-up" }
 log = "0.4"
+tokio = { version = "1", features = ["macros", "rt", "rt-multi-thread", "time"] }
